Some Edit

+++ Task1-Crud/app/Http/Controllers/ArticaleController.php
@@ -7,6 +7,7 @@ use App\Http\Requests\UpdateArticaleRequest;
 use App\Http\Services\Media;
 use App\Models\Articale;
 use App\Models\Category;
+use Illuminate\Support\Facades\Artisan;
 
 class ArticaleController extends Controller
 {
@@ -37,29 +38,33 @@ class ArticaleController extends Controller
     }
 
 
-    public function edit(Articale $articale)
+    public function edit($id)
     {
+        $articale = Articale::find($id);
         $categories = Category::all();
         return view("Articales/update",compact('articale','categories'));
     }
 
 
-    public function update(UpdateArticaleRequest $request, Articale $articale)
+    public function update(UpdateArticaleRequest $request,int $id)
     {
         $data = $request->except('_token','_method','image');
         if($request->hasFile('image')){
             $dir=public_path('images/articales');
            $newimage= Media::upload($dir,$request->file('image'));
+           $articale = Articale::find($id);
             Media::delete($dir."//$articale->image");
             $data['image']=$newimage;
         }
-        return $this->redirectBack($articale->where('id',$articale->id)->update($data),'articale.index');
+        return $this->redirectBack(Articale::where('id',$id)->update($data),'articale.index');
     }
 
-    public function destroy(Articale $articale)
+    public function destroy(int $id)
     {
         #delete image from server
-        dd($articale);
+        $articale = Articale::find($id);
+        if(!$articale)
+            return abort(404);
         Media::delete(public_path("images/articales/$articale->image"));
         $articale->delete();
         return redirect()->back()->with('success','Articale Deleted Successfully');
